+++ app/src/main/java/com/example/graduationproject/authentication/signup/model/UserData.kt
@@ -1,0 +1,5 @@
+package com.example.graduationproject.authentication.signup.model
+
+data class UserData(val userName:String, val phoneNumber:String, val password:String, val imageUri:String, val userId:String){
+
+}
